requirements.txt
streamlit==1.22.0
pandas==1.5.3
matplotlib==3.7.0
smbus2==0.5.0
bmp280==0.1.4